Move user to seperate file, implement authorize in seperate file to prevent cookies error from client imports
