+++ back-end/routes/api.php
@@ -17,15 +17,12 @@ Route::get('/health', [HealthController::class, 'index']);
 
 Route::prefix('auth')->group(function (): void {
     Route::post('/login', [AuthenticatedSessionController::class, 'store'])
-        ->middleware('guest')
         ->name('login');
 
     Route::post('/forgot-password', [PasswordResetLinkController::class, 'store'])
-        ->middleware('guest')
         ->name('password.email');
 
     Route::post('/reset-password', [NewPasswordController::class, 'store'])
-        ->middleware('guest')
         ->name('password.store');
 
     Route::post('/logout', [AuthenticatedSessionController::class, 'destroy'])
