+++ front-end/src/lib/auth.ts
@@ -1,38 +1,11 @@
 import axios from "@/lib/axios";
-import { ApiResponseError } from "@/lib/errors";
+import { authorize } from "@/lib/signInApi";
+import { User } from "@/types/user";
 import { JWT } from "@auth/core/jwt";
 import Credentials from "@auth/core/providers/credentials";
-import { AxiosError } from "axios";
-import NextAuth, { NextAuthConfig, Session } from "next-auth";
-import { cookies } from "next/headers";
+import NextAuth, { Session } from "next-auth";
 
-interface ApiLoginResponse {
-  user: ApiUser;
-}
-
-export interface ApiUser {
-  UserID: number;
-  UserFirstName: string;
-  UserLastName: string;
-  email: string;
-  DateHired: Date;
-  UserRoleID: number;
-  RoleName: string;
-  created_at: string | null;
-  updated_at: string | null;
-}
-
-export interface User {
-  id: string;
-  firstName: string;
-  lastName: string;
-  email: string;
-  dateHired: Date;
-  roleId: number;
-  roleName: string;
-}
-
-export const authOptions: NextAuthConfig = {
+export const { handlers, signIn, signOut, auth } = NextAuth({
   providers: [
     Credentials({
       credentials: {
@@ -43,54 +16,7 @@ export const authOptions: NextAuthConfig = {
           type: "password",
         },
       },
-      authorize: async (credentials) => {
-        try {
-          const loginResponse = await axios.post<ApiLoginResponse>(
-            "/auth/login",
-            credentials,
-          );
-
-          const responseCookies = loginResponse.headers["set-cookie"];
-          if (!responseCookies) {
-            throw new Error("No session cookies in response");
-          }
-
-          // Parse and set each cookie in the browser to allow future requests
-          for (const cookie of responseCookies) {
-            axios.defaults.headers.common["Cookie"] =
-              responseCookies.join("; ");
-
-            const [cookieName, ...rest] = cookie.split("=");
-            const cookieValue = rest.join("=").split(";")[0];
-
-            (await cookies()).set(cookieName, cookieValue, {
-              sameSite: "lax",
-            });
-          }
-
-          const apiUser = loginResponse.data.user;
-
-          return {
-            id: apiUser.UserID.toString(),
-            firstName: apiUser.UserFirstName,
-            lastName: apiUser.UserLastName,
-            email: apiUser.email,
-            dateHired: new Date(apiUser.DateHired),
-            roleId: apiUser.UserRoleID,
-            roleName: apiUser.RoleName,
-          };
-        } catch (error) {
-          if (
-            !(error instanceof AxiosError) ||
-            error.response?.status !== 422
-          ) {
-            return null; // Return null to display the default error message
-          }
-
-          // In this case the error contains user-friendly error messages
-          throw new ApiResponseError(error.response.data.message);
-        }
-      },
+      authorize,
     }),
   ],
   events: {
@@ -111,6 +37,4 @@ export const authOptions: NextAuthConfig = {
     },
   },
   basePath: "/auth",
-};
-
-export const { handlers, signIn, signOut, auth } = NextAuth(authOptions);
+});
