+++ front-end/src/types/user.ts
@@ -1,0 +1,9 @@
+export interface User {
+  id: string;
+  firstName: string;
+  lastName: string;
+  email: string;
+  dateHired: Date;
+  roleId: number;
+  roleName: string;
+}
